requirements.txt
# 보안: CVE-2024-24762(ReDoS) 대응 — 0.109.1 이상 권장
fastapi>=0.109.1
# CVE-2020-7694/7695는 0.11.7 미만 이슈. 장기적으로 최신 안정 버전 유지
uvicorn[standard]>=0.24.0
# 2.9.x 최신으로 버그·호환성 수정 반영
psycopg2-binary>=2.9.11
# CVE-2024-3219 등 보안·호환성 개선 반영
kafka-python>=2.0.3
